Support of RPC scheduling
To perform operations while device is disconnected (e.g., USB is off),
RPCs need to be scheduled with certain delay before disconnection
happens.

While device is disconnected, as long as the snippet is still running,
the scheduled RPCs should be able to be executed.

The RPC result could be retrieved as cached events once device get back
online.