Add support for long arrays as snippet parameters
Specifying a snippet that takes a long array as a parameter, currently fails because there is no explicit conversion specified for the long array type. As a result, the "Try any custom converter provided" else block gets executed and fails because it tries to convert a JSONArray to a JSONObject.

Using an integer array as an alternative is also problematic, because the size of a Python integer can exceed the size of a Java integer.

This change adds explicit support for long arrays.