Add support for long arrays as snippet parameters (#105)
Specifying a snippet that takes a long array as a parameter, currently fails because there is no explicit conversion specified for the long array type. Using an integer array as an alternative is also problematic, because the size of a Python integer can exceed the size of a Java integer.